+++ spec/requests/projects_controller_spec.rb
@@ -59,6 +59,64 @@ RSpec.describe ProjectsController, type: :request do
     end
   end
 
+  describe "#openers" do
+    context "when the month and year are missing" do
+      it "returns a 404" do
+        get "/projects/openers"
+        expect(response).to have_http_status(:not_found)
+      end
+    end
+
+    context "when the year is missing" do
+      it "returns a 404" do
+        get "/projects/openers/1"
+        expect(response).to have_http_status(:not_found)
+      end
+    end
+
+    context "when the month isn't in the scope 1..12" do
+      it "returns a 404" do
+        get "/projects/openers/13/2022"
+        expect(response).to have_http_status(:not_found)
+      end
+    end
+
+    context "when the year isn't in the scope 2000-2499" do
+      it "returns a 404" do
+        get "/projects/openers/12/2555"
+        expect(response).to have_http_status(:not_found)
+      end
+    end
+
+    context "when the month and year are present and in scope" do
+      it "shows a page title with the month & year" do
+        get "/projects/openers/1/2022"
+        expect(response.body).to include("Academies opening in January 2022")
+      end
+
+      # TODO: page currently shows all conversion projects
+      it "returns project details in table form" do
+        conversion_project = create(:conversion_project)
+        get "/projects/openers/1/2022"
+        expect(response.body).to include(
+          conversion_project.establishment.name,
+          conversion_project.urn.to_s,
+          conversion_project.incoming_trust.name,
+          conversion_project.incoming_trust.ukprn
+        )
+      end
+
+      xit "only returns projects whose confirmed conversion date is in that month & year"
+
+      context "when there are no academies opening in that month & year" do
+        xit "shows a helpful message" do
+          get "/projects/openers/01/2022"
+          expect(response.body).to include("There are currently no schools expected to become academies in January 2022")
+        end
+      end
+    end
+  end
+
   private def unset_api_data
     # Nillify any project data that comes from the API in order to best test the
     # real scenario, where loaded project data will not have associated API data until