+++ CHANGELOG.md
@@ -23,6 +23,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - the width of the application has been increased from 960px to 1024px when
   viewed on desktop class device.
 
+### Added
+
+- add a Project Openers table view
+
 ## [Release 15][release-15]
 
 ### Changed